requirements.txt
numpy
Pillow
pydub
